requirements.txt
anyio==3.6.2
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.4
Brlapi==0.8.3
camelot-py==0.9.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
comm==0.1.2
command-not-found==0.3
cryptography==39.0.0
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
fqdn==1.5.1
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.20.2
ipython==8.9.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==8.10.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.68
openpyxl==3.1.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.16.2
pdfminer.six==20221105
pdfplumber==0.7.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.21.1
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==2.12.1
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-json-logger==2.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.0
reportlab==3.6.8
requests==2.25.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
systemd-python==234
tabula-py==2.6.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.2.0
urllib3==1.26.5
wadllib==1.3.6
Wand==0.6.11
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
xdg==5
xkit==0.0.0
zipp==1.0.0
